pyproject.toml
[project]
name = "zenith-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "amqp==5.2.0",
    "asgiref==3.8.1",
    "async-timeout==4.0.3",
    "attrs==23.2.0",
    "billiard==4.2.0",
    "celery==5.4.0",
    "certifi==2024.7.4",
    "cffi==1.16.0",
    "channels==4.1.0",
    "channels-redis==4.2.0",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "cron-descriptor==1.4.3",
    "cryptography==43.0.0",
    "defusedxml==0.7.1",
    "diff-match-patch==20230430",
    "dj-database-url==2.2.0",
    "django==5.0.7",
    "django-celery-beat==2.6.0",
    "django-celery-results==2.5.1",
    "django-cors-headers==4.4.0",
    "django-debug-toolbar==4.4.6",
    "django-extensions==3.2.3",
    "django-filter==24.2",
    "django-import-export==4.1.1",
    "django-redis==5.4.0",
    "django-templated-mail==1.1.1",
    "django-timezone-field==7.0",
    "django-unfold==0.33.0",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "djoser==2.2.3",
    "drf-nested-routers==0.94.1",
    "drf-spectacular==0.27.2",
    "exceptiongroup==1.2.2",
    "gunicorn==22.0.0",
    "h11==0.14.0",
    "h2==4.1.0",
    "hpack==4.0.0",
    "hypercorn==0.17.3",
    "hyperframe==6.0.1",
    "idna==3.7",
    "inflection==0.5.1",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2023.12.1",
    "kombu==5.3.7",
    "markdown==3.6",
    "msgpack==1.0.8",
    "oauthlib==3.2.2",
    "packaging==24.1",
    "paho-mqtt==2.1.0",
    "pillow==10.4.0",
    "priority==2.0.0",
    "prompt-toolkit==3.0.47",
    "psycopg==3.2.1",
    "psycopg-binary==3.2.1",
    "pycparser==2.22",
    "pyjwt==2.8.0",
    "python-crontab==3.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python3-openid==3.2.0",
    "pyyaml==6.0.1",
    "redis==5.0.7",
    "referencing==0.35.1",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "rpds-py==0.19.1",
    "six==1.16.0",
    "social-auth-app-django==5.4.2",
    "social-auth-core==4.5.4",
    "sqlparse==0.5.1",
    "tablib==3.5.0",
    "taskgroup==0.0.0a4",
    "tomli==2.0.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.1",
    "uritemplate==4.1.1",
    "urllib3==2.2.2",
    "uvicorn==0.30.3",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "whitenoise==6.7.0",
    "wsproto==1.2.0",
]
